LPS-55479 move to internal package
