+++ modules/portal/portal-cache-ehcache/src/com/liferay/portal/cache/ehcache/internal/EhcacheCacheManagerListenerAdapter.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.cache.ehcache;
+package com.liferay.portal.cache.ehcache.internal;
 
 import com.liferay.portal.kernel.cache.CacheManagerListener;
 import com.liferay.portal.kernel.cache.PortalCacheException;
