+++ modules/portal/portal-cache-ehcache/src/com/liferay/portal/cache/ehcache/internal/EhcacheConstants.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.cache.ehcache;
+package com.liferay.portal.cache.ehcache.internal;
 
 /**
  * @author Tina Tian
