+++ modules/portal/portal-cache-ehcache/src/com/liferay/portal/cache/ehcache/internal/EhcacheUnwrapUtil.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.cache.ehcache;
+package com.liferay.portal.cache.ehcache.internal;
 
 import com.liferay.portal.kernel.cache.PortalCache;
 import com.liferay.portal.kernel.cache.PortalCacheWrapper;
