+++ modules/portal/portal-cache-ehcache/test/unit/com/liferay/portal/cache/ehcache/internal/EhcachePortalCacheTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.cache.ehcache;
+package com.liferay.portal.cache.ehcache.internal;
 
 import com.liferay.portal.cache.test.MockPortalCacheManager;
 import com.liferay.portal.cache.test.TestCacheListener;
